Add authentication endpoint and handling.
Add paragraph about development.

Allowing the `Location` header to be read from the browser makes for a
nicer experience as we can send a HTML page *and* parse the intended
redirect target in the web application.  Therefore, if a user were to
open the raw endpoint they'd still be redirected as we want them to.